test: テストの修正

+++ api/server/user/lib/array/convert_test.go
@@ -27,7 +27,7 @@ func TestConvertStrings(t *testing.T) {
 			Expected: []string{"1", "2", "3"},
 		},
 		"ng_other_type": {
-			Input:    []int{1, 2, 3},
+			Input:    []float64{1.0, 2.0, 3.0},
 			Expected: []string{},
 		},
 	}
